AUTOREVIEW minor - Pin the dtype of the cross-block dedup test
The test built its payload with np.array([10, 11, ...]), whose dtype is the
platform default integer: int64 on Linux and macOS, but int32 on Windows. The
target raw_data is int64 everywhere, so on Windows the new test compared the
int64 dataset contents against an int32 expectation of a different shape and
failed. Pin dtype=np.int64 so the test is platform-independent.

+++ tests/test_backend.py
@@ -911,15 +911,16 @@ def test_rewrite_dataset_dedups_across_blocks(vfile):
     """
     create_base_dataset(vfile.f, "x", data=np.empty(0, dtype=np.int64), chunks=(2,))
     # Four blocks of one chunk each (a chunk is 16 bytes == max_bytes): block 2
-    # duplicates block 1 and block 3 duplicates block 0
-    data = np.array([10, 11, 20, 21, 20, 21, 10, 11])
+    # duplicates block 1 and block 3 duplicates block 0. Pin the dtype: the default
+    # integer is 32-bit on Windows, and the target raw_data is int64 everywhere.
+    data = np.array([10, 11, 20, 21, 20, 21, 10, 11], dtype=np.int64)
     sc = rewrite_dataset(vfile.f, "x", data, chunks=(2,), max_bytes=16)
 
     raw_data, hash_table = _raw_data_hashtable(vfile, "x")
     # Only two of the four chunks are original; the other two are not written again
     assert raw_data.shape == (4,)
     assert hash_table.attrs["largest_index"] == 2
-    assert_equal(raw_data[:], np.array([10, 11, 20, 21]))
+    assert_equal(raw_data[:], np.array([10, 11, 20, 21], dtype=np.int64))
     assert_equal(sc[()], data)
     # Block 1 landed at offset 2 and block 2 deduplicates onto it, not onto offset 0
     assert_equal(sc.slab_offsets, np.array([0, 2, 2, 0], dtype=sc.slab_offsets.dtype))
